requirements.txt
Django>=5.0
djangorestframework
djoser
djangorestframework-simplejwt
